Add proper error handling on SDK load
The promise returned from `sdk.loaded` will now get rejected if
there is an error communicating with the backend server to fetch
user information.

This does require the developer to set up a proper promise handler
chain using `sdk.loaded().catch()` otherwise an Unhandled Rejection
error will be printed to the console if an error occurs. The docs
have been updated to reflect this.