Adding the Missing Twitch Features Helpers
isBitsEnabled :: If this flag is true, Bits in Extensions features will work in your extension on the current channel

isChatEnabled :: If this flag is true, you can send a chat message to the current channel using Send Extension Chat Message (subject to the authentication requirements documented for that endpoint).

isSubscriptionStatusAvailable :: If this flag is true, your extension has the ability to get the subscription status of identity-linked viewers from both the helper in the window.Twitch.ext.viewer.subscriptionStatus object and via the Twitch API.

onChanged :: This function enables you to receive real-time updates to changes of the features object. If this callback is invoked, you should re-check the window.Twitch.ext.features object for a change to any feature flag your extension cares about. The callback is called with an array of feature flags which were updated.

Incrementing package version.